requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
filelock==3.15.4
fsspec==2024.6.1
idna==3.7
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
openai-whisper==20231117
pydub==0.25.1
regex==2024.7.24
requests==2.32.3
sympy==1.13.1
tiktoken==0.7.0
torch==2.2.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
